refactor: BoardResponseDto DTO에는 비즈니스 로직이 들어가면 안되므로, 서비스 계층으로 비즈니스 로직 옮김
