+++ src/main/java/com/example/yun/domain/board/Board.java
@@ -35,7 +35,6 @@ public class Board extends BaseEntity {
     public void updateTitle(String title) {
         this.title = title;
     }
-
     public void updateContent(String content) {
         this.content = content;
     }
