+++ src/main/java/com/example/yun/service/impl/BoardServiceImpl.java
@@ -4,7 +4,7 @@ import com.example.yun.domain.board.Board;
 import com.example.yun.dto.BoardResponseDto;
 import com.example.yun.repository.BoardRepository;
 import com.example.yun.repository.querydsl.BoardQueryRepository;
-import com.example.yun.service.BoardMessage;
+import com.example.yun.exception.BoardMessage;
 import com.example.yun.service.BoardService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -54,7 +54,7 @@ public class BoardServiceImpl implements BoardService {
 
         log.info("boards = {}", boards);
 
-        return BoardResponseDto.responseDtosCreate(boards);
+        return responseDtosCreate(boards);
     }
 
     /**
@@ -140,7 +140,7 @@ public class BoardServiceImpl implements BoardService {
     public List<BoardResponseDto> boardAllSearchBySort() {
         List<Board> boards = boardQueryRepository.boardAllSearchBySort();
 
-        return BoardResponseDto.responseDtosCreate(boards);
+        return responseDtosCreate(boards);
     }
 
     /**
@@ -152,7 +152,7 @@ public class BoardServiceImpl implements BoardService {
     public List<BoardResponseDto> boardAllSearchByKeyword(String keyword) {
         List<Board> boards = boardQueryRepository.boardSearchByKeyword(keyword);
 
-        return BoardResponseDto.responseDtosCreate(boards);
+        return responseDtosCreate(boards);
     }
 
     /**
@@ -172,5 +172,10 @@ public class BoardServiceImpl implements BoardService {
      */
     private static Board getBoard(Optional<Board> board) {
         return board.get();
+    }
+
+    public static List<BoardResponseDto> responseDtosCreate(List<Board> boards) {
+        return boards.stream().map(BoardResponseDto::from)
+                .collect(toList());
     }
 }
